+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.0] - 2020-12-07
+
 ### Added
  - Added access to the CloudBrigadierManager from Brigadier-enabled command managers
  - Added parameter injectors (cloud-annotations)
@@ -32,6 +34,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Fix Bukkit alias command suggestions without Brigadier
  - Fix Bukkit command alias registration when using Brigadier
  - Fixed a bug where providing valid input for an argument caused cloud to no longer make suggestions
+ - Detect and throw an exception for ambiguous nodes in more cases
+ - CloudBrigadierManager no longer forgets the command sender
 
 ## [1.1.0] - 2020-10-24
 
@@ -59,8 +63,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  
 ### Fixed
  - Fix arguments with no required children not being executors (cloud-brigadier)
- - Detect and throw an exception for ambiguous nodes in more cases
- - CloudBrigadierManager no longer forgets the command sender
 
 ## [1.0.2] - 2020-10-18